Restructuring of model-part of ptoject.
