+++ src/app/features/components/capture/running-events/running-events.component.spec.ts
@@ -2,7 +2,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { RunningEventsComponent } from './running-events.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { IRunningEventsBusinessLogicModel, RunningEventsBusinessLogicModel } from '../model/capture/business-logic-model/running-events-business-logic-model';
+import { IRunningEventsBusinessLogicModel, RunningEventsBusinessLogicModel } from '../../../models/capture/business-logic-model/running-events-business-logic-model';
 import { IPersistedEvent } from '../../../../shared/classes/db/time-series-db';
 import { RunningEventsPersistence } from "../../../../shared/classes/db/running-events-db";
 import { Logger } from '../../../../shared/services/logging/logger';
