+++ src/app/features/models/capture/business-logic-model/capture-business-logic-model.ts
@@ -1,9 +1,9 @@
-import { Logger } from "../../../../../../shared/services/logging/logger";
-import { IMetaDataPersistence } from "../../../../../../shared/classes/db/metadata-db";
+import { Logger } from "../../../../shared/services/logging/logger";
+import { IMetaDataPersistence } from "../../../../shared/classes/db/metadata-db";
 import { IRepositoryBusinessLogicModel } from "./repository-navigation-business-logic-model";
 import { RepositoryBusinessLogicModel } from "./repository-navigation-business-logic-model";
 import { IRunningEventsBusinessLogicModel } from "./running-events-business-logic-model";
-import { CurrentEventProcessingBusinessLogicModel } from "./current-event-business-logic-model";
+import { CurrentEventProcessingBusinessLogicModel } from "./current-event-business-logic-model/current-event-business-logic-model";
 
 
 export interface ICaptureBusinessLogicModel {
