+++ src/app/features/models/capture/capture-common-interfaces.ts
