+++ src/app/features/models/capture/ui-model/capture-ui-model.ts
@@ -1,10 +1,10 @@
 
-import { Logger } from '../../../../../../shared/services/logging/logger';
+import { Logger } from '../../../../shared/services/logging/logger';
 import { CaptureBusinessLogicModel, 
     ICaptureBusinessLogicModel} from '../business-logic-model/capture-business-logic-model';
 import { IRepositoryNavigationUIModel } from './repository-navigation-ui-model';
 import { IRunningEventsUIModel } from './running-events-ui-model';
-import { CurrentEventProcessingUIModel } from './current-event-ui-model';
+import { CurrentEventProcessingUIModel } from './current-event-ui-model/current-event-ui-model';
 
 
 // UI model for the capture component
